refactor(operations): reuse the cluster normalization's compDef resolver in upgrade precheck
Extract resolveCompDefinitionNServiceVersion and its helpers from
controllers/apps/cluster into pkg/controller/component as the exported
ResolveCompDefinitionNServiceVersion, and replace the duplicated
pattern/serviceVersion resolution in the upgrade ops precheck with it.
The deterministic no-match failure is now signaled by the errors.Is-able
ErrCompDefinitionNServiceVersionUnmatched sentinel, which the ops layer
classifies as a fatal error.
