+++ controllers/apps/cluster/transformer_cluster_normalization_test.go
@@ -203,49 +203,49 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("full match", func() {
 			By("with definition v1.0 and service version v0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition v1.1 and service version v0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition v2.0 and service version v0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition v1.0 and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v1.1 and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v2.0 and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v3.0 and service version v2")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v3.0"), testapps.ServiceVersion("v3"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v3.0"), testapps.ServiceVersion("v3"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
@@ -254,28 +254,28 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("w/o service version", func() {
 			By("with definition v1.0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), "")
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v1.1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.1"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v2.0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2.0"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v3.0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v3.0"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v3.0"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
@@ -284,35 +284,35 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("prefix match definition", func() {
 			By("with definition prefix and service version v0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition prefix and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition prefix and service version v2")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v3"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, testapps.ServiceVersion("v3"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r5", "r5")
 
 			By("with definition v1 prefix and service version v0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition v2 prefix and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
@@ -321,21 +321,21 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("prefix match definition and w/o service version", func() {
 			By("with definition prefix")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, "")
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefinitionName, "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r5", "r5")
 
 			By("with definition v1 prefix")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v2 prefix")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v2"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
@@ -344,35 +344,35 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("regular expression match definition", func() {
 			By("with definition exact regex and service version 1")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v2.0"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v2.0"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition exact regex and service version v2")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v2.0"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v2.0"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition exact regex and service version v3")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v3.0"), testapps.ServiceVersion("v3"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithExactRegex("v3.0"), testapps.ServiceVersion("v3"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r5", "r5")
 
 			By("with definition v1 fuzzy regex and service version v0")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v1"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v1"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r4", "r4")
 
 			By("with definition v2 fuzzy regex and service version v1")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v2"), testapps.ServiceVersion("v2"))
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v2"), testapps.ServiceVersion("v2"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
@@ -381,21 +381,21 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("regular expression match definition and w/o service version", func() {
 			By("with definition regex")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, "^"+testapps.CompDefinitionName, "")
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, "^"+testapps.CompDefinitionName, "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v3.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v3")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r5", "r5")
 
 			By("with definition v1 regex")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v1"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v1"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.1")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
 			updateNCheckCompDefinitionImages(compDef, resolvedServiceVersion, "r3", "r2")
 
 			By("with definition v2 regex")
-			compDef, resolvedServiceVersion, err = resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v2"), "")
+			compDef, resolvedServiceVersion, err = component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefNameWithFuzzyRegex("v2"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v2.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
@@ -404,7 +404,7 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("match from definition", func() {
 			By("with definition v1.0 and service version v0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v0"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v0"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v0")))
@@ -451,7 +451,7 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 			})).Should(Succeed())
 
 			By("with definition v4.0 and service version v3")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v4.0"), testapps.ServiceVersion("v4"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v4.0"), testapps.ServiceVersion("v4"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v4.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v4")))
@@ -649,7 +649,7 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("full match", func() {
 			By("with definition v1.0 and service version v0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v1"))
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), testapps.ServiceVersion("v1"))
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v1")))
@@ -658,7 +658,7 @@ var _ = Describe("resolve CompDefinition and ServiceVersion", func() {
 
 		It("w/o service version", func() {
 			By("with definition v1.0")
-			compDef, resolvedServiceVersion, err := resolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), "")
+			compDef, resolvedServiceVersion, err := component.ResolveCompDefinitionNServiceVersion(testCtx.Ctx, testCtx.Cli, testapps.CompDefName("v1.0"), "")
 			Expect(err).Should(Succeed())
 			Expect(compDef.Name).Should(Equal(testapps.CompDefName("v1.0")))
 			Expect(resolvedServiceVersion).Should(Equal(testapps.ServiceVersion("v2")))
